Simplify TSC calculations
